config: update to use new 4 machines config

+++ dsm-scripts/config_memdev.sh
@@ -1,12 +1,17 @@
 mode=$1
 memNumaNode=5
+size=16 # 16GB shared memory
 devName="/dev/shm/ivshmem-$USER"
 
-rm -rf $devName
-echo "Old Shared Memory Removed"
+if [ "$mode" = "cxl-new" ]; then
+	rm -rf $devName
+	echo "Old Shared Memory Removed"
+	# sudo daxctl reconfigure-device --mode=system-ram dax0.0 --force
+	numactl --membind=$memNumaNode dd if=/dev/zero of=$devName bs=1G count=$size
+	echo "New Shared Memory (on NUMA $memNumaNode) Malloced"
+fi
 
 if [ "$mode" = "cxl" ]; then
-	# sudo daxctl reconfigure-device --mode=system-ram dax0.0 --force
-	numactl --membind=$memNumaNode dd if=/dev/zero of=$devName bs=1G count=4
-	echo "New Shared Memory (on NUMA $memNumaNode) Malloced"
+	dd if=/dev/zero of=$devName bs=1 count=1024 conv=notrunc
+	echo "Set first 1KB of Shared Memory (on NUMA $memNumaNode) to 0"
 fi
